// src/app/models/consumerData.ts
export interface IConsumerFormData {
    firstname: string;
    lastname: string;
    dob: string;
    businessname: string;
    pandetails: string;
    email: string;
    phone: string;
    website: string;
    businessoverview: string;
    validity: string;
    agentname: string;
    agentid: number;
    businesscategory: string;
    businesstype: string;
    businessturnover: number;
    capitalinvested: number;
    totalemployees: number;
    businessage: number;
}